Moved extention function to utils
