+++ dependencies.gradle
@@ -2,7 +2,7 @@ ext.versions = [
     min_sdk: 16,
     compile_sdk: 29,
     build_tools: "29.0.0",
-    publish_version: "3.3.0",
+    publish_version: "3.3.1",
     publish_version_code: 262
 ]
